Implemented/added putUserQualifications as a replacement for postUserQualifications with varied functionality based on new requirements.
